[PIN-96] Improved exponential cost/reward algorithms (#34)
* new exponential cost algo

* reorganize typescript cloud functions into modules

* add `cost` field to `Pin`'s

* add new reward function based on `Pin` cost
