+++ functions/src/types.ts
@@ -1,0 +1,17 @@
+import { GeoPoint } from "firebase-admin/firestore";
+
+export enum PinType {
+  TEXT = "TEXT",
+  IMAGE = "IMAGE",
+}
+
+export type Pin = {
+  caption: string;
+  textContent?: string;
+  type: PinType;
+  location: GeoPoint;
+  authorUID: string;
+  timestamp: Date;
+  finds: number;
+  cost: number;
+};
